Fixed prison system!
Woop Woop! I fixed the commit I made early accidentally. You will
probably notice I also reenabled the ability to jail people from the
admin menu, I needed this for development purposes to test it. I can
redisable it if you want before this is merged but I feel it poses as a
useful tool for the moderators. The performance change with the new code
should be negligible but it does involve updating the player prison bail
every second. I'm shattered atm so I havent been able to think of an
algorithm using percentages that isn't broken like the previous one but
I promise I will eventually.  I feel it is better to have this way as it
does mean you can actually part way pay off your bail as opposed to
having to pay the whole thing.